requirements.txt
beautifulsoup4==4.9.3
decorator==5.0.9
greenlet==1.1.0
Pillow==8.2.0
psycopg2-binary==2.9.1
pytesseract==0.3.7
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.19
SQLAlchemy-Utils==0.37.7
urllib3==1.26.6
validators==0.18.2
